PahamSatuMenit/requirements.txt
openai==0.27.0
ratelimit==2.2.1
python-dotenv==1.0.1
pandas==2.2.3
numpy==2.0.2
nltk==3.9.1 
bert_score==0.3.13  
rouge_score==0.1.2